requirements.txt
python==3.10
PyTorch==1.4.0